Optimise core UI test scope
