+++ packages/core-ui/src/tests/app/import-export-workspace.test.js
@@ -1,8 +1,7 @@
 import { GenericDataTable } from '@anywaydata/core/data_formats/generic-data-table.js';
 import { jest } from '@jest/globals';
-import { fireEvent, waitFor } from '@testing-library/dom';
+import { fireEvent } from '@testing-library/dom';
 import { JSDOM } from 'jsdom';
-import * as importExportWorkspaceExports from '../../../js/gui_components/app/import-export-workspace/index.js';
 import { createImportExportWorkspaceComponent } from '../../../js/gui_components/app/import-export-workspace/index.js';
 
 function createTable(rowCount = 3) {
@@ -37,6 +36,12 @@ function getCopyTextButton(documentObj) {
 
 function getCopyTextButtonLabel(documentObj) {
   return documentObj.querySelector('[data-role="text-preview-editor-root"] [data-role="copy-text-label"]');
+}
+
+async function flushAsyncWork(iterations = 4) {
+  for (let index = 0; index < iterations; index += 1) {
+    await Promise.resolve();
+  }
 }
 
 function createHarness({ props = {}, services = {}, navigatorPlatform } = {}) {
@@ -101,14 +106,6 @@ describe('ImportExportWorkspace', () => {
     jest.restoreAllMocks();
     delete global.document;
     delete global.window;
-  });
-
-  test('public barrel is component-factory-only', () => {
-    expect(importExportWorkspaceExports.createImportExportWorkspaceComponent).toBe(
-      createImportExportWorkspaceComponent
-    );
-    expect(importExportWorkspaceExports.ImportExportWorkspaceController).toBeUndefined();
-    expect(importExportWorkspaceExports.ImportExportWorkspaceView).toBeUndefined();
   });
 
   test('mounts real controls, normalizes preview row count, and renders format previews without legacy controls', () => {
@@ -216,7 +213,8 @@ describe('ImportExportWorkspace', () => {
 
     fireEvent.click(importButton);
 
-    await waitFor(() => expect(importer.setGridFromGenericDataTable).toHaveBeenCalledTimes(1));
+    await flushAsyncWork();
+    expect(importer.setGridFromGenericDataTable).toHaveBeenCalledTimes(1);
     expect(getPreviewTextArea(documentObj).value).toBe('csv:rows:2');
     expect(documentObj.querySelector('#import-progress-status').textContent).toBe('Import complete.');
     expect(component.getState().previewTextDirty).toBe(false);
@@ -258,14 +256,13 @@ describe('ImportExportWorkspace', () => {
 
     fireEvent.click(documentObj.querySelector('#filedownload'));
 
-    await waitFor(() =>
-      expect(downloadService.downloadText).toHaveBeenCalledWith('export.csv', 'full:csv', {
-        exportEncodingSettings: {
-          lineEnding: 'lf',
-          includeBom: true,
-        },
-      })
-    );
+    await flushAsyncWork();
+    expect(downloadService.downloadText).toHaveBeenCalledWith('export.csv', 'full:csv', {
+      exportEncodingSettings: {
+        lineEnding: 'lf',
+        includeBom: true,
+      },
+    });
     expect(documentObj.querySelector('#export-progress-status').textContent).toBe('Download started.');
 
     component.destroy();
@@ -291,12 +288,13 @@ describe('ImportExportWorkspace', () => {
     });
     fireEvent.change(fileInput);
 
-    await waitFor(() => expect(fileReadService.readText).toHaveBeenCalledTimes(1));
+    await flushAsyncWork();
+    expect(fileReadService.readText).toHaveBeenCalledTimes(1);
     expect(importer.setImportSettings).toHaveBeenCalledWith({
       trimInput: false,
       trimInputFieldsCsv: '',
     });
-    await waitFor(() => expect(importer.setGridFromGenericDataTable).toHaveBeenCalledTimes(1));
+    expect(importer.setGridFromGenericDataTable).toHaveBeenCalledTimes(1);
     expect(getPreviewTextArea(documentObj).value).toBe('csv:rows:1');
     expect(documentObj.querySelector('#import-progress-status').textContent).toBe('Import complete.');
 
@@ -317,16 +315,15 @@ describe('ImportExportWorkspace', () => {
 
     fireEvent.click(documentObj.querySelector('[data-role="clipboard-import-button"]'));
 
-    await waitFor(() => expect(clipboardService.readText).toHaveBeenCalledTimes(1));
+    await flushAsyncWork();
+    expect(clipboardService.readText).toHaveBeenCalledTimes(1);
     expect(importer.setImportSettings).toHaveBeenCalledWith({
       trimInput: false,
       trimInputFieldsCsv: '',
     });
-    await waitFor(() => expect(importer.importText).toHaveBeenCalledWith('csv', 'Name,Role\nAda,Engineer'));
-    await waitFor(() => expect(getPreviewTextArea(documentObj).value).toBe('csv:rows:1'));
-    await waitFor(() =>
-      expect(documentObj.querySelector('#import-progress-status').textContent).toBe('Import complete.')
-    );
+    expect(importer.importText).toHaveBeenCalledWith('csv', 'Name,Role\nAda,Engineer');
+    expect(getPreviewTextArea(documentObj).value).toBe('csv:rows:1');
+    expect(documentObj.querySelector('#import-progress-status').textContent).toBe('Import complete.');
 
     component.destroy();
     dom.window.close();
@@ -403,7 +400,8 @@ describe('ImportExportWorkspace', () => {
       readerInstance.emit('progress', { loaded: 5, total: 10, lengthComputable: true });
       readerInstance.emit('load', { target: { result: 'Name,Role\nAda,Engineer' } });
 
-      await waitFor(() => expect(importer.setGridFromGenericDataTable).toHaveBeenCalledTimes(1));
+      await flushAsyncWork();
+      expect(importer.setGridFromGenericDataTable).toHaveBeenCalledTimes(1);
       expect(getPreviewTextArea(documentObj).value).toBe('csv:rows:1');
       expect(documentObj.querySelector('#import-progress-status').textContent).toBe('Import complete.');
     } finally {
@@ -442,17 +440,17 @@ describe('ImportExportWorkspace', () => {
       trimInput: false,
       trimInputFieldsCsv: '',
     });
-    await waitFor(() => expect(importer.importText).toHaveBeenCalledWith('csv', textArea.value));
+    await flushAsyncWork();
+    expect(importer.importText).toHaveBeenCalledWith('csv', textArea.value);
     expect(component.getState().importBusy).toBe(false);
     expect(documentObj.querySelector('#import-progress-status').textContent).toBe('Import complete.');
 
     importer.importText.mockRejectedValueOnce(new Error('boom'));
     fireEvent.click(importButton);
 
-    await waitFor(() =>
-      expect(documentObj.querySelector('#import-progress-status').textContent).toBe(
-        'Import failed. Check file format/options.'
-      )
+    await flushAsyncWork();
+    expect(documentObj.querySelector('#import-progress-status').textContent).toBe(
+      'Import failed. Check file format/options.'
     );
     expect(component.getState().importBusy).toBe(false);
 
@@ -534,12 +532,11 @@ describe('ImportExportWorkspace', () => {
     });
 
     fireEvent.click(documentObj.querySelector('[data-role="clipboard-import-button"]'));
-    await waitFor(() =>
-      expect(importer.setImportSettings).toHaveBeenCalledWith({
-        trimInput: true,
-        trimInputFieldsCsv: 'Role',
-      })
-    );
+    await flushAsyncWork();
+    expect(importer.setImportSettings).toHaveBeenCalledWith({
+      trimInput: true,
+      trimInputFieldsCsv: 'Role',
+    });
 
     const fileInput = documentObj.querySelector('#csvinput');
     Object.defineProperty(fileInput, 'files', {
@@ -548,12 +545,11 @@ describe('ImportExportWorkspace', () => {
     });
     fireEvent.change(fileInput);
 
-    await waitFor(() =>
-      expect(importer.setImportSettings).toHaveBeenLastCalledWith({
-        trimInput: true,
-        trimInputFieldsCsv: 'Role',
-      })
-    );
+    await flushAsyncWork();
+    expect(importer.setImportSettings).toHaveBeenLastCalledWith({
+      trimInput: true,
+      trimInputFieldsCsv: 'Role',
+    });
 
     component.destroy();
     dom.window.close();
